Separating functional tests into modules
